requirements.txt
pygame>=2.0.0
pyinstaller>=5.0.0
pystray>=0.19.0
Pillow>=9.0.0
win10toast>=0.9
tkinterdnd2>=0.3.0

